+++ scanpipe/tests/test_commands.py
@@ -20,6 +20,7 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/nexB/scancode.io for support and download.
 
+import datetime
 import tempfile
 from io import StringIO
 from pathlib import Path
@@ -119,16 +120,16 @@ class ScanPipeManagementCommandTest(TestCase):
     def test_scanpipe_management_command_create_project_inputs(self):
         out = StringIO()
 
-        options = ["--input", "non-existing"]
+        options = ["--input-file", "non-existing"]
         expected = "non-existing not found or not a file"
         with self.assertRaisesMessage(CommandError, expected):
             call_command("create-project", "my_project", *options)
 
         parent_path = Path(__file__).parent
         options = [
-            "--input",
+            "--input-file",
             str(parent_path / "test_commands.py"),
-            "--input",
+            "--input-file",
             str(parent_path / "test_models.py"),
         ]
         call_command("create-project", "my_project", *options, stdout=out)
@@ -158,13 +159,15 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn(f"Pipeline {pipeline} run in progress...", out.getvalue())
         self.assertIn("successfully executed on project my_project", out.getvalue())
 
-    def test_scanpipe_management_command_add_input(self):
+    def test_scanpipe_management_command_add_input_file(self):
         out = StringIO()
 
         project = Project.objects.create(name="my_project")
         parent_path = Path(__file__).parent
         options = [
+            "--input-file",
             str(parent_path / "test_commands.py"),
+            "--input-file",
             str(parent_path / "test_models.py"),
         ]
 
@@ -178,10 +181,22 @@ class ScanPipeManagementCommandTest(TestCase):
         expected = sorted(["test_commands.py", "test_models.py"])
         self.assertEqual(expected, sorted(project.input_files))
 
-        options = ["--project", project.name, "non-existing.py"]
+        options = ["--project", project.name, "--input-file", "non-existing.py"]
         expected = "non-existing.py not found or not a file"
         with self.assertRaisesMessage(CommandError, expected):
             call_command("add-input", *options, stdout=out)
+
+    def test_scanpipe_management_command_add_input_url(self):
+        out = StringIO()
+
+        project = Project.objects.create(name="my_project")
+        parent_path = Path(__file__).parent
+        options = [
+            "--input-file",
+            str(parent_path / "test_commands.py"),
+            "--input-file",
+            str(parent_path / "test_models.py"),
+        ]
 
     def test_scanpipe_management_command_add_pipeline(self):
         out = StringIO()
@@ -221,7 +236,7 @@ class ScanPipeManagementCommandTest(TestCase):
         options = ["--project", project.name, "--no-color"]
         out = StringIO()
         call_command("show-pipeline", *options, stdout=out)
-        expected = " [ ] docker\n" " [ ] root_filesystems\n"
+        expected = " [NOT_STARTED] docker\n" " [NOT_STARTED] root_filesystems\n"
         self.assertEqual(expected, out.getvalue())
 
         project.runs.filter(pipeline_name=pipeline_names[0]).update(task_exitcode=0)
@@ -273,7 +288,7 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn("- CodebaseResource: 0", output)
         self.assertIn("- DiscoveredPackage: 0", output)
         self.assertIn("- ProjectError: 0", output)
-        self.assertIn("[ ] docker", output)
+        self.assertIn("[NOT_STARTED] docker", output)
 
         run.task_start_date = timezone.now()
         run.log = (
@@ -288,6 +303,15 @@ class ScanPipeManagementCommandTest(TestCase):
         self.assertIn("[RUNNING] docker", output)
         for line in run.log.split("\n"):
             self.assertIn(line, output)
+
+        run.task_end_date = run.task_start_date + datetime.timedelta(0, 42)
+        run.task_exitcode = 0
+        run.save()
+        out = StringIO()
+        call_command("status", *options, stdout=out)
+        output = out.getvalue()
+        expected = f"[SUCCESS] docker (executed in {run.execution_time} seconds)"
+        self.assertIn(expected, output)
 
     def test_scanpipe_management_command_output(self):
         project = Project.objects.create(name="my_project")